requirements.txt
fastapi~=0.115.5
openai~=1.55.0
pydantic~=2.10.1
uvicorn~=0.32.1
python-dotenv